Used srtmasc2bin to converts some files. Divide by 3 the size of the file.

+++ src/lib_hmap/hmap.h
@@ -95,7 +95,7 @@ class HMap
 	//! name of the map
 	string m_name;
 	//! storage area
-	valarray<double> m_v;
+	valarray<int16_t> m_v;
 	//! full size
 	size_t m_size;
 	//! nb of rows
